requirements.txt
anthropic>=0.18.0
PyMuPDF>=1.23.5
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
flake8>=7.0.0